Adapt start-feder8-studio.sh to enable security for Linux based systems

+++ remote-installation/separate-scripts/start-feder8-studio.sh
@@ -5,6 +5,13 @@ VERSION=2.0.3
 TAG=$VERSION
 CURRENT_DIRECTORY=$(pwd)
 
+if command -v systemctl &> /dev/null
+then
+    LINUX=true
+else
+    LINUX=false
+fi
+
 read -p "Enter the Therapeutic Area of choice. Enter honeur/phederation/esfurn/athena [honeur]: " HONEUR_THERAPEUTIC_AREA
 while [[ "$HONEUR_THERAPEUTIC_AREA" != "honeur" && "$HONEUR_THERAPEUTIC_AREA" != "phederation" && "$HONEUR_THERAPEUTIC_AREA" != "esfurn" && "$HONEUR_THERAPEUTIC_AREA" != "athena" && "$HONEUR_THERAPEUTIC_AREA" != "" ]]; do
     echo "Enter \"honeur\", \"phederation\", \"esfurn\", \"athena\" or empty for default \"honeur\" value"
@@ -39,28 +46,30 @@ while [[ "$HONEUR_CLI_SECRET" == "" ]]; do
     read -p "Enter the CLI Secret: " HONEUR_CLI_SECRET
 done
 
-read -p "Enter the folder containing the certificates [$CURRENT_DIRECTORY/certificates]: " CERTIFICATE_FOLDER
-CERTIFICATE_FOLDER=${CERTIFICATE_FOLDER:-$CURRENT_DIRECTORY/certificates}
-if [ -d "$CERTIFICATE_FOLDER" ]; then
-    mkdir -p $CERTIFICATE_FOLDER/$HONEUR_THERAPEUTIC_AREA-studio
-    rm -rf $CERTIFICATE_FOLDER/$HONEUR_THERAPEUTIC_AREA-studio/*
-    if [ -f "$CERTIFICATE_FOLDER/$HONEUR_THERAPEUTIC_AREA-studio-client-key.pem" ]; then
-        cp -v $CERTIFICATE_FOLDER/$HONEUR_THERAPEUTIC_AREA-studio-client-key.pem $CERTIFICATE_FOLDER/$HONEUR_THERAPEUTIC_AREA-studio/key.pem
-        cp -v $CERTIFICATE_FOLDER/$HONEUR_THERAPEUTIC_AREA-studio-client-cert.pem $CERTIFICATE_FOLDER/$HONEUR_THERAPEUTIC_AREA-studio/cert.pem
-        cp -v $CERTIFICATE_FOLDER/ca.pem $CERTIFICATE_FOLDER/$HONEUR_THERAPEUTIC_AREA-studio/ca.pem
-        chmod -v 0400 $CERTIFICATE_FOLDER/$HONEUR_THERAPEUTIC_AREA-studio/key.pem
-        chmod -v 0444 $CERTIFICATE_FOLDER/$HONEUR_THERAPEUTIC_AREA-studio/cert.pem
-    elif [ -f "$CERTIFICATE_FOLDER/key.pem" ]; then
-        cp -v $CERTIFICATE_FOLDER/key.pem $CERTIFICATE_FOLDER/$HONEUR_THERAPEUTIC_AREA-studio/key.pem
-        cp -v $CERTIFICATE_FOLDER/cert.pem $CERTIFICATE_FOLDER/$HONEUR_THERAPEUTIC_AREA-studio/cert.pem
-        cp -v $CERTIFICATE_FOLDER/ca.pem $CERTIFICATE_FOLDER/$HONEUR_THERAPEUTIC_AREA-studio/ca.pem
+if $LINUX; then
+    read -p "Enter the folder containing the certificates [$CURRENT_DIRECTORY/certificates]: " CERTIFICATE_FOLDER
+    CERTIFICATE_FOLDER=${CERTIFICATE_FOLDER:-$CURRENT_DIRECTORY/certificates}
+    if [ -d "$CERTIFICATE_FOLDER" ]; then
+        mkdir -p $CERTIFICATE_FOLDER/$HONEUR_THERAPEUTIC_AREA-studio
+        rm -rf $CERTIFICATE_FOLDER/$HONEUR_THERAPEUTIC_AREA-studio/*
+        if [ -f "$CERTIFICATE_FOLDER/$HONEUR_THERAPEUTIC_AREA-studio-client-key.pem" ]; then
+            cp -v $CERTIFICATE_FOLDER/$HONEUR_THERAPEUTIC_AREA-studio-client-key.pem $CERTIFICATE_FOLDER/$HONEUR_THERAPEUTIC_AREA-studio/key.pem
+            cp -v $CERTIFICATE_FOLDER/$HONEUR_THERAPEUTIC_AREA-studio-client-cert.pem $CERTIFICATE_FOLDER/$HONEUR_THERAPEUTIC_AREA-studio/cert.pem
+            cp -v $CERTIFICATE_FOLDER/ca.pem $CERTIFICATE_FOLDER/$HONEUR_THERAPEUTIC_AREA-studio/ca.pem
+            chmod -v 0400 $CERTIFICATE_FOLDER/$HONEUR_THERAPEUTIC_AREA-studio/key.pem
+            chmod -v 0444 $CERTIFICATE_FOLDER/$HONEUR_THERAPEUTIC_AREA-studio/cert.pem
+        elif [ -f "$CERTIFICATE_FOLDER/key.pem" ]; then
+            cp -v $CERTIFICATE_FOLDER/key.pem $CERTIFICATE_FOLDER/$HONEUR_THERAPEUTIC_AREA-studio/key.pem
+            cp -v $CERTIFICATE_FOLDER/cert.pem $CERTIFICATE_FOLDER/$HONEUR_THERAPEUTIC_AREA-studio/cert.pem
+            cp -v $CERTIFICATE_FOLDER/ca.pem $CERTIFICATE_FOLDER/$HONEUR_THERAPEUTIC_AREA-studio/ca.pem
+        else
+            echo "Warning: '$CERTIFICATE_FOLDER' doesn't contain a client certificate for ${HONEUR_THERAPEUTIC_AREA_UPPERCASE} Studio.  Abort."
+            exit
+        fi
     else
-        echo "Warning: '$CERTIFICATE_FOLDER' doesn't contain a client certificate for ${HONEUR_THERAPEUTIC_AREA_UPPERCASE} Studio.  Abort."
+        echo "Warning: '$CERTIFICATE_FOLDER' NOT found.  Abort."
         exit
     fi
-else
-    echo "Warning: '$CERTIFICATE_FOLDER' NOT found.  Abort."
-    exit
 fi
 
 read -p 'Enter the FQDN(Fully Qualified Domain Name eg. www.example.com) or public IP address(eg. 125.24.44.18) of the host machine. Use localhost to for testing [localhost]: ' HONEUR_HOST_MACHINE
@@ -107,8 +116,10 @@ echo "AUTHENTICATION_METHOD=$HONEUR_SECURITY_METHOD" >> honeur-studio.env
 echo "HONEUR_THERAPEUTIC_AREA=$HONEUR_THERAPEUTIC_AREA" >> honeur-studio.env
 echo "HONEUR_THERAPEUTIC_AREA_URL=${HONEUR_THERAPEUTIC_AREA_URL}" >> honeur-studio.env
 echo "HONEUR_THERAPEUTIC_AREA_UPPERCASE=${HONEUR_THERAPEUTIC_AREA_UPPERCASE}" >> honeur-studio.env
-echo "PROXY_DOCKER_URL=https://172.17.0.1:2376" >> honeur-studio.env
-echo "PROXY_DOCKER_CERT_PATH=/home/certs" >> honeur-studio.env
+if $LINUX; then
+    echo "PROXY_DOCKER_URL=https://172.17.0.1:2376" >> honeur-studio.env
+    echo "PROXY_DOCKER_CERT_PATH=/home/certs" >> honeur-studio.env
+fi
 
 if [ "$HONEUR_SECURITY_METHOD" = "jdbc" ]; then
     #JDBC
@@ -171,20 +182,37 @@ docker network connect $HONEUR_THERAPEUTIC_AREA-studio-net $HONEUR_THERAPEUTIC_A
 
 
 echo "Run $HONEUR_THERAPEUTIC_AREA/$HONEUR_THERAPEUTIC_AREA-studio:$TAG container. This could take a while..."
-docker run \
---name "$HONEUR_THERAPEUTIC_AREA-studio" \
---restart on-failure:5 \
---security-opt no-new-privileges \
---env-file honeur-studio.env \
--v "shared:/var/lib/shared:ro" \
--v "/var/run/docker.sock:/var/run/docker.sock" \
--m "1g" \
---cpus "2" \
---pids-limit 100 \
---cpu-shares 1024 \
---ulimit nofile=1024:1024 \
--d \
-$HONEUR_THERAPEUTIC_AREA_URL/$HONEUR_THERAPEUTIC_AREA/$HONEUR_THERAPEUTIC_AREA-studio:$TAG shinyproxy
+if $LINUX; then
+  docker run \
+  --name "$HONEUR_THERAPEUTIC_AREA-studio" \
+  --restart on-failure:5 \
+  --security-opt no-new-privileges \
+  --env-file honeur-studio.env \
+  -v "shared:/var/lib/shared:ro" \
+  -v "$CERTIFICATE_FOLDER/honeur-studio:/home/certs" \
+  -m "1g" \
+  --cpus "2" \
+  --pids-limit 100 \
+  --cpu-shares 1024 \
+  --ulimit nofile=1024:1024 \
+  -d \
+  $HONEUR_THERAPEUTIC_AREA_URL/$HONEUR_THERAPEUTIC_AREA/$HONEUR_THERAPEUTIC_AREA-studio:$TAG shinyproxy
+else
+  docker run \
+  --name "$HONEUR_THERAPEUTIC_AREA-studio" \
+  --restart on-failure:5 \
+  --security-opt no-new-privileges \
+  --env-file honeur-studio.env \
+  -v "shared:/var/lib/shared:ro" \
+  -v "/var/run/docker.sock:/var/run/docker.sock" \
+  -m "1g" \
+  --cpus "2" \
+  --pids-limit 100 \
+  --cpu-shares 1024 \
+  --ulimit nofile=1024:1024 \
+  -d \
+  $HONEUR_THERAPEUTIC_AREA_URL/$HONEUR_THERAPEUTIC_AREA/$HONEUR_THERAPEUTIC_AREA-studio:$TAG shinyproxy
+fi
 
 echo Connect $HONEUR_THERAPEUTIC_AREA-studio to $HONEUR_THERAPEUTIC_AREA-net network
 docker network connect $HONEUR_THERAPEUTIC_AREA-net $HONEUR_THERAPEUTIC_AREA-studio > /dev/null 2>&1
